Refactor termFilter into a shared directive
