+++ app/budget/directives/budgetScenarioToolbar/budgetScenarioToolbar.html
@@ -58,9 +58,9 @@
 	</div>
 
 	<div class="budget-scenario-toolbar__tool-container">
-		<term-filter active-terms="state.selectedBudgetScenario.terms"
-		             selected-budget-scenario="state.selectedBudgetScenario">
-		</term-filter>
+			<term-filter active-terms="state.selectedBudgetScenario.terms"
+			             select-term="selectBudgetScenarioTerm">
+			</term-filter>
 
 		<span class="budget-scenario-toolbar__tool glyphicon glyphicon-download-alt" data-toggle="dropdown"></span>
 		<span class="budget-scenario-toolbar__tool glyphicon glyphicon-print"></span>
