+++ app/shared/directives/termFilter/termFilter.html
@@ -19,11 +19,11 @@
 			</div>
 
 			<div class="term-filter__term-list">
-				<div ng-click="selectBudgetScenarioTerm(term)"
+				<div ng-click="clickTerm(term)"
 				     ng-repeat="term in allTerms"
 				     class="term-filter__term">
 					<ipa-checkbox is-checked="(activeTerms.indexOf(term) > -1)"
-					              click-action="selectBudgetScenarioTerm(term)">
+					              click-action="clickTerm(term)">
 					</ipa-checkbox>
 					<div class="term-filter__term-text">
 						{{ getTermName(term) }}
